requirements.txt
flask
line-bot-sdk>=3.0.0
gunicorn
gspread
oauth2client
python-dotenv